% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cox_Mknockoff.R
\name{Cox_Mknockoff}
\alias{Cox_Mknockoff}
\title{Cox_Mknockoff}
\usage{
Cox_Mknockoff(
  X,
  Y,
  M,
  COV,
  k,
  fdr,
  penalty = c("MCP", "SCAD", "lasso"),
  stat = c("beta", "lambda"),
  knockoff = c(TRUE, FALSE),
  topN = NULL,
  verbose = TRUE
)
}
\arguments{
\item{X}{X is the vector of exposure variable.}

\item{Y}{Y is the vector of outcome variable.}

\item{M}{M is the matrix of mediators.}

\item{COV}{COV is the vector/matrix of covariate(s).}

\item{k}{k is an integer to adjust the number of mediators in SIS step.}

\item{fdr}{fdr is the targeted false discovery rate that we want to control. A value between 0 and 1.}

\item{penalty}{penalty is the type of penalty we use.}

\item{stat}{stat is the type of knockoff statistic we use.}

\item{knockoff}{knockoff is a boolean value to control whether we use knockoff method or not.}

\item{topN}{topN is an integer. If given, then we directly choose topN mediators in SIS step based on the statistics.}

\item{verbose}{if TRUE, then details will be printed}
}
\value{
a list of outcomes containing the selected mediators and the corresponding results.
}
\description{
Apply knockoff methods into high-dimensional mediation analysis.
}
