% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_stat.R
\name{knockoff_stat}
\alias{knockoff_stat}
\title{knockoff_stat}
\usage{
knockoff_stat(beta, lambda)
}
\arguments{
\item{beta}{beta is the vector of coefficients corresponding to lambda.}

\item{lambda}{lambda is the vector of penalties.}
}
\value{
a vector
}
\description{
Calculate the knockoff statistic based on the result of penalized regression.
}
