% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_surv.R
\name{knockoff_surv}
\alias{knockoff_surv}
\title{knockoff_surv}
\usage{
knockoff_surv(
  X,
  Y,
  M_SIS,
  COV,
  penalty = c("MCP", "SCAD", "lasso"),
  fdr,
  stat = c("beta", "lambda")
)
}
\arguments{
\item{X}{X is the vector of exposure variable.}

\item{Y}{Y is the vector of outcome variable.}

\item{M_SIS}{M_SIS is the matrix of mediators after sure independence screening.}

\item{COV}{COV is the vector/matrix of covariate(s).}

\item{penalty}{penalty is the type of penalty we use.}

\item{fdr}{fdr is the targeted false discovery rate that we want to control. A value between 0 and 1.}

\item{stat}{stat is the type of knockoff statistic we use.}
}
\value{
a vector of selected mediators.
}
\description{
Knockoff method for survival outcome.
}
