% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R
\name{proportion}
\alias{proportion}
\title{proportion}
\usage{
proportion(v, W)
}
\arguments{
\item{v}{v is a given number}

\item{W}{W is the vector of knockoff statistic.}
}
\value{
a value
}
\description{
Calculate the proportion of a given num, whicberh will be used in calculating the threshold.
}
