% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_alpha.R
\name{sis_alpha}
\alias{sis_alpha}
\title{sis_alpha}
\usage{
sis_alpha(X, M, COV)
}
\arguments{
\item{X}{X is the vector of exposure variable.}

\item{M}{M is the matrix of mediators.}

\item{COV}{COV is the vector/matrix of covariate(s).}
}
\value{
a list containing variance, estimation and pval for alpha
}
\description{
Sure independence screening based on X->M pathway.
}
