% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_beta.R
\name{sis_beta}
\alias{sis_beta}
\title{sis_beta}
\usage{
sis_beta(X, M, Y, COV)
}
\arguments{
\item{X}{X is the vector of exposure variable.}

\item{M}{M is the matrix of mediators.}

\item{Y}{Y is the vector of outcome variable.}

\item{COV}{COV is the vector/matrix of covariate(s).}
}
\value{
a list containing variance, estimation and pval for beta
}
\description{
Sure independence screening based on M->Y pathway.
}
