% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\title{threshold}
\usage{
threshold(W, fdr)
}
\arguments{
\item{W}{W is the vector of knockoff statistic.}

\item{fdr}{fdr is the targeted false discovery rate that we want to control. A value between 0 and 1.}
}
\value{
a value
}
\description{
Get the threshold of knockoff statistic.
}
